@@ -10,47 +10,54 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@microflash/remark-figure-caption": "^1.0.0",
+		"@portaljs/remark-callouts": "^1.0.5",
 		"@sveltejs/adapter-static": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/enhanced-img": "^0.3.10",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@tailwindcss/forms": "^0.5.3",
 		"@tailwindcss/typography": "^0.5.9",
-		"@types/eslint": "8.56.0",
-		"@typescript-eslint/eslint-plugin": "^5.59.2",
-		"@typescript-eslint/parser": "^5.59.2",
+		"@threlte/core": "^8.0.1",
+		"@threlte/extras": "^9.0.1",
+		"@types/d3": "^7.4.3",
+		"@types/eslint": "^9.6.0",
+		"@types/three": "^0.172.0",
 		"@vcarl/remark-headings": "^0.1.0",
 		"autoprefixer": "^10.4.14",
-		"eslint": "^8.40.0",
-		"eslint-config-prettier": "^8.8.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"heroicons-svelte": "^2.0.0",
-		"mdsvex": "^0.11.0",
-		"mdsvex-relative-images": "^1.0.3",
+		"camera-controls": "^2.8.3",
+		"d3": "^7.9.0",
+		"eslint": "^9.0.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^2.45.1",
+		"heroicons-svelte": "^2.0.2",
+		"mdsvex": "^0.12.3",
+		"mdsvex-enhanced-images": "^0.2.3",
 		"postcss": "^8.4.23",
-		"postcss-load-config": "^4.0.1",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"rehype-accessible-emojis": "^0.3.2",
-		"rehype-autolink-headings": "^6.1.1",
-		"rehype-slug": "^5.1.0",
-		"remark-gemoji": "^7.0.1",
-		"remark-gfm": "^3.0.1",
-		"remark-toc": "^8.0.1",
-		"remark-unwrap-images": "^3.0.1",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"svelte-preprocess": "^5.0.3",
-		"svelte-youtube-embed": "^0.2.1",
-		"sveltekit-embed": "^0.0.12",
+		"rehype-autolink-headings": "^7.1.0",
+		"rehype-github-alert": "^1.0.0",
+		"rehype-slug": "^6.0.0",
+		"rehype-unwrap-images": "^1.0.0",
+		"remark-gfm": "^4.0.0",
+		"remark-toc": "^9.0.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"svelte-preprocess": "^6.0.2",
+		"svelte-scrolling": "^1.5.1",
+		"svelte-youtube-embed": "^0.4.4",
+		"sveltekit-embed": "^0.0.18",
 		"tailwindcss": "^3.3.1",
+		"three": "^0.172.0",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.0",
+		"typescript": "^5.5.0",
+		"typescript-eslint": "^8.0.0",
+		"vite": "^6.0.11",
 		"vite-plugin-glsl": "^1.1.2"
 	},
 	"type": "module",
 	"dependencies": {
-		"eslint-plugin-svelte": "^2.32.4"
+		"eslint-plugin-svelte": "^2.45.1"
 	}
 }
