@@ -1,9 +1,13 @@
 <script lang="ts">
-	export let description: string;
+	interface Props {
+		description: string | null | undefined;
+	}
+
+	let { description }: Props = $props();
 </script>
 
 {#if description}
-	<p class="text-sm italic text-eigengrau-900/70 dark:text-cosmiclatte-50/50">
+	<p class="text-xs italic text-eigengrau-900/50 dark:text-cosmiclatte-50/50">
 		{description}
 	</p>
 {/if}
