@@ -1,12 +1,14 @@
 <script lang="ts">
-	import type { SvelteComponentTyped } from 'svelte/internal';
+	//import type { SvelteComponentTyped } from 'svelte/internal';
 	import PageHead from '$lib/components/PageHead.svelte';
 	import ArticleTitle from '$lib/components/ArticleTitle.svelte';
 	import ArticleMeta from '$lib/components/ArticleMeta.svelte';
 	import ArticleDescription from '$lib/components/ArticleDescription.svelte';
-	export let data;
+	let { data } = $props();
 	type C = $$Generic<typeof SvelteComponentTyped<any, any, any>>;
-	$: component = data.component as unknown as C;
+	let component = $derived(data.component as unknown as C);
+
+	const SvelteComponent = $derived(component);
 </script>
 
 <PageHead
@@ -30,15 +32,29 @@
 			<ul class="article-tags">
 				{#each data.frontmatter.tags as tag}
 					<li class="">
-						<span class="tag">{tag}</span>
+						<a href="/topics/{encodeURIComponent(tag)}" class="tag">{tag}</a>
 					</li>
 				{/each}
 			</ul>
 		{/if}
+		<!-- {#if data.frontmatter.headings && data.frontmatter.headings.length > 5}
+			<div class="article-headings text-left sticky mt-3 bg-slate-500/50 py-2">
+				<ul class="font-normal text-xs">
+					{#each data.frontmatter.headings as heading}
+						<li class="heading heading-{heading.depth}">
+							<span class=""
+								><a href="#{heading.id}">{heading.value}</a></span
+							>
+						</li>
+					{/each}
+				</ul>
+			</div>
+		{/if} -->
 	</div>
 	<hr class="mb-4" />
+	<!-- <div>{JSON.stringify(data, null, 2)} -- aa</div> -->
 	<div class="content prose prose-stone dark:prose-invert">
-		<svelte:component this={component} />
+		<SvelteComponent />
 	</div>
 </article>
 
@@ -46,9 +62,21 @@
 	.meta {
 		/* @apply mt-16; */
 		@apply flex items-baseline;
-		@apply lg:absolute lg:w-[15ch] lg:-ml-[18ch] lg:text-right;
+		@apply lg:absolute lg:w-[15ch] lg:-ml-[18ch] lg:text-right xl:w-[25ch] xl:-ml-[28ch];
 		@apply lg:flex-col;
 		@apply text-stone-400 font-mono text-xs font-light;
+		.heading {
+			@apply text-xs;
+		}
+		.heading-2{
+			@apply pl-2;
+		}
+		.heading-3{
+			@apply pl-4;
+		}
+		.heading-4{
+			@apply pl-5;
+		}
 	}
 	.date-info {
 		@apply order-last flex-1 flex justify-end gap-1 items-baseline;
