@@ -1,14 +1,28 @@
 import { defineMDSveXConfig as defineConfig } from 'mdsvex';
 
-import remarkGemoji from 'remark-gemoji';
-import relativeImages from 'mdsvex-relative-images';
-import remarkFigureCaption from '@microflash/remark-figure-caption';
-import remarkUnwrapImages from 'remark-unwrap-images';
+/** REMARK plugins - (work on markdown AST) */
+//import plugin from 'remark-github-beta-blockquote-admonitions'
+//import { remarkAlert } from 'remark-github-blockquote-alert'
+//import remarkCallouts from '@portaljs/remark-callouts'
+
+// transform and process the images to <picture> with srcset
+import { enhancedImages } from 'mdsvex-enhanced-images';
+// A remark plugin to add metadata about headings to the parsed output.
 import remarkHeadings from '@vcarl/remark-headings';
+// github flavored markdown
 import remarkGfm from 'remark-gfm';
+// generates a TOC
 import remarkToc from 'remark-toc';
-import autolinkHeadings from 'rehype-autolink-headings';
+
+/** REHYPE plugins - (work on outputed HTML) */
+//remove the wrapping paragraph (<p>) for images (<img>)
+import rehypeUnwrapImages from 'rehype-unwrap-images';
+//  transform an image with alt text to a figure with caption
+import rehypeGithubAlert from 'rehype-github-alert';
+// plugin to add ids to headings
 import slugPlugin from 'rehype-slug';
+// generates a link for each heading who has an id
+import autolinkHeadings from 'rehype-autolink-headings';
 import rehypeAccessibleEmojis from 'rehype-accessible-emojis';
 
 const config = defineConfig({
@@ -18,18 +32,12 @@ const config = defineConfig({
 		dashes: 'oldschool'
 	},
 
-	remarkPlugins: [
-		relativeImages,
-		remarkUnwrapImages,
-		remarkFigureCaption,
-		remarkGfm,
-		headings,
-		remarkGemoji,
-		[remarkToc, { tight: true }]
-	],
+	remarkPlugins: [enhancedImages, remarkGfm, headings, [remarkToc, { tight: true }]],
 	rehypePlugins: [
 		rehypeAccessibleEmojis,
+		rehypeUnwrapImages,
 		slugPlugin,
+		rehypeGithubAlert,
 		[
 			autolinkHeadings,
 			{
