@@ -19,19 +19,25 @@
 		find me elsewhere:
 		<ul>
 			<li>
-				<a href="https://twitter.com/0gust1" rel="">twitter</a>
-			</li>
-			<li>
 				<a href="https://merveilles.town/@0gust1">merveilles.town</a>
 			</li>
 			<li>
 				<a href="https://github.com/0gust1">github</a>
+			</li>
+			<li>
+				<a href="https://bsky.app/profile/0gust1.bsky.social">bluesky</a>
+			</li>
+			<li>
+				<a href="https://twitter.com/0gust1" class="cancel">twitter</a>
 			</li>
 		</ul>
 	</div>
 </div>
 
 <style lang="postcss">
+	.cancel {
+		@apply line-through decoration-solid decoration-red-500;
+	}
 	ul {
 		@apply inline-flex;
 	}
