@@ -1,7 +1,8 @@
 import { mdsvex } from 'mdsvex';
 import mdsvexConfig from './mdsvex.config.js';
-import preprocess from 'svelte-preprocess';
+import { sveltePreprocess } from 'svelte-preprocess'
 import adapter from '@sveltejs/adapter-static';
+
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
@@ -12,7 +13,7 @@ const config = {
 	// for more information about preprocessors
 	preprocess: [
 		vitePreprocess(),
-		preprocess({
+		sveltePreprocess({
 			postcss: true
 		}),
 		mdsvex(mdsvexConfig)
